GLT-2582 don't override locationBarcode
