+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/AbstractLibrary.java
@@ -203,11 +203,6 @@ public abstract class AbstractLibrary extends AbstractBoxable implements Library
   @Transient
   private List<FileAttachment> pendingAttachmentDeletions;
 
-  private boolean distributed;
-  @Temporal(TemporalType.DATE)
-  private Date distributionDate;
-  private String distributionRecipient;
-
   @Override
   public Boxable.EntityType getEntityType() {
     return Boxable.EntityType.LIBRARY;
@@ -757,36 +752,6 @@ public abstract class AbstractLibrary extends AbstractBoxable implements Library
   @Override
   public void setPendingAttachmentDeletions(List<FileAttachment> pendingAttachmentDeletions) {
     this.pendingAttachmentDeletions = pendingAttachmentDeletions;
-  }
-
-  @Override
-  public boolean isDistributed() {
-    return distributed;
-  }
-
-  @Override
-  public void setDistributed(boolean distributed) {
-    this.distributed = distributed;
-  }
-
-  @Override
-  public Date getDistributionDate() {
-    return distributionDate;
-  }
-
-  @Override
-  public void setDistributionDate(Date distributionDate) {
-    this.distributionDate = distributionDate;
-  }
-
-  @Override
-  public String getDistributionRecipient() {
-    return distributionRecipient;
-  }
-
-  @Override
-  public void setDistributionRecipient(String distributionRecipient) {
-    this.distributionRecipient = distributionRecipient;
   }
 
 }
