+++ sqlstore/src/main/resources/db/migration/V8234__make_ld_and_pools_distributable.sql
@@ -1,0 +1,7 @@
+ALTER TABLE LibraryDilution ADD COLUMN distributed BOOLEAN NOT NULL DEFAULT FALSE;
+ALTER TABLE LibraryDilution ADD COLUMN distributionDate DATE DEFAULT NULL;
+ALTER TABLE LibraryDilution ADD COLUMN distributionRecipient VARCHAR(250) DEFAULT NULL;
+
+ALTER TABLE Pool ADD COLUMN distributed BOOLEAN NOT NULL DEFAULT FALSE;
+ALTER TABLE Pool ADD COLUMN distributionDate DATE DEFAULT NULL;
+ALTER TABLE Pool ADD COLUMN distributionRecipient VARCHAR(250) DEFAULT NULL;
